Compare the class nibble as a byte in the illusion temple class test

+++ tests/MUnique.OpenMU.Tests/IllusionTempleCharacterClassTests.cs
@@ -72,6 +72,6 @@ public class IllusionTempleCharacterClassTests
     {
         var converted = characterClassNumber.ToIllusionTempleCharacterClass();
 
-        Assert.That(converted & 0xF0, Is.EqualTo(MasterClass));
+        Assert.That((byte)(converted & 0xF0), Is.EqualTo(MasterClass));
     }
 }
